fix: 768 이상의 해상도에서 랜딩페이지 좌우 패딩 오류 (#808)
* fix: mainPage 768px 이상 해상도에서 제대로 작동하지 않던 오류 해결

* style: lint 적용
