+++ client/src/GlobalStyle.ts
@@ -142,8 +142,6 @@ export const GlobalStyle = css`
       sans-serif;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
-    max-width: 768px;
-    margin: 0 auto;
   }
 
   scrollbar-width: none; /* Firefox */
